Tests/LibWeb/Text/expected/HTML/details-name.txt

details0=✗ details1=✗ details2=✗ details3=✗ details4=✗
details0=✓ details1=✗ details2=✗ details3=✗ details4=✗
details0=✗ details1=✓ details2=✗ details3=✗ details4=✗
details0=✗ details1=✗ details2=✓ details3=✗ details4=✗
details0=✗ details1=✗ details2=✓ details3=✓ details4=✗
details0=✗ details1=✗ details2=✓ details3=✓ details4=✓
details0=✓ details1=✗ details2=✗ details3=✓ details4=✓
